move sdputils  out from initAudio. move moudules out from Amps.
