+++ js/amp/amp.js
@@ -1,23 +1,25 @@
-
+import Impulses from './impulses.js';
+import { Convolver } from './convolver.js';
+import { Visualization } from './visualization.js';
 
 // INITS
+
 var audioPlayer, input2;
 var demoSampleURLs = [
-  "assets/audio/Guitar_DI_Track.mp3",
-  "assets/audio/LasseMagoDI.mp3",
-  "assets/audio/RawPRRI.mp3",
-  "assets/audio/Di-Guitar.mp3",
-  "assets/audio/NarcosynthesisDI.mp3",
-  "assets/audio/BlackSabbathNIB_rythmDI.mp3",
-  "assets/audio/BlackSabbathNIBLead_DI.mp3",
-  "assets/audio/BasketCase Greenday riff DI.mp3",
-  "assets/audio/InBloomNirvanaRiff1DI.mp3",
-  "assets/audio/Muse1Solo.mp3",
-  "assets/audio/Muse2Rythm.mp3"
+  'assets/audio/Guitar_DI_Track.mp3',
+  'assets/audio/LasseMagoDI.mp3',
+  'assets/audio/RawPRRI.mp3',
+  'assets/audio/Di-Guitar.mp3',
+  'assets/audio/NarcosynthesisDI.mp3',
+  'assets/audio/BlackSabbathNIB_rythmDI.mp3',
+  'assets/audio/BlackSabbathNIBLead_DI.mp3',
+  'assets/audio/BasketCase Greenday riff DI.mp3',
+  'assets/audio/InBloomNirvanaRiff1DI.mp3',
+  'assets/audio/Muse1Solo.mp3',
+  'assets/audio/Muse2Rythm.mp3'
 ];
 
-
-function gotStream() {
+export function gotStream(audioContext) {
     // Create an AudioNode from the stream.
     audioPlayer = document.getElementById('player');
     try {
@@ -25,24 +27,48 @@ function gotStream() {
         // WebComponent polyfill that wraps the HTML audio
         // element into something that cannot be used with
         // createMediaElementSource. We use ShadowDOMPolyfill.unwrap(...)
-        // to get the "real" HTML audio element
+        // to get the 'real' HTML audio element
         audioPlayer = ShadowDOMPolyfill.unwrap(audioPlayer);
-    } catch(e) {
-        console.log("ShadowDOMPolyfill undefined, running native Web Component code");
+    }
+    catch (e) {
+        console.log('ShadowDOMPolyfill undefined, running native Web Component code');
     }
 
-    
-    if(input2 === undefined) {
-        input2 = audioContext.createMediaElementSource(audioPlayer);
+    if (input2 === undefined) {
+        // input2 = audioContext.createMediaElementSource(audioPlayer);
     }
 
     var input = audioContext.createMediaStreamSource(window.stream);
-    audioInput = convertToMono(input);
+    var audioInput = convertToMono(audioContext, input);
 
     createAmp(audioContext, audioInput, input2);
     console.log('AMP CREATED')
 }
 
+function toggleGuitarInput(event) {
+    var button = document.querySelector('#toggleGuitarIn');
+
+    if (!guitarPluggedIn) {
+        guitarInput.connect(amp.input);
+        button.innerHTML = 'Guitar input: <span style="color:green;">ACTIVATED</span>, click to toggle on/off!';
+        button.classList.remove('pulse');
+    } else {
+        guitarInput.disconnect();
+        button.innerHTML = 'Guitar input: <span style="color:red;">NOT ACTIVATED</span>, click to toggle on/off!';
+        button.classList.add('pulse');
+    }
+    guitarPluggedIn = !guitarPluggedIn;
+}
+
+function convertToMono(audioContext, input) {
+    var splitter = audioContext.createChannelSplitter(2);
+    var merger = audioContext.createChannelMerger(2);
+
+    input.connect(splitter);
+    splitter.connect(merger, 0, 0);
+    splitter.connect(merger, 0, 1);
+    return merger;
+};
 
 function changeDemoSample(val) {
     console.log(val);
@@ -63,16 +89,17 @@ function createAmp(context, input1, input2) {
 
     // create quadrafuzz
     amp = new Amp(context);
+    window.amp = amp;
     analyzerAtInput = context.createAnalyser();
     amp.input.connect(analyzerAtInput);
 
     // build graph
-    if(guitarPluggedIn) {
+    if (guitarPluggedIn) {
         guitarInput.connect(amp.input);
     }
 
     // connect audio player to amp for previewing presets
-    input2.connect(amp.input);
+    // input2.connect(amp.input);
 
     // output, add an analyser at the end
     analyzerAtOutput = context.createAnalyser();
@@ -85,30 +112,15 @@ function createAmp(context, input1, input2) {
     initVisualizations();
 }
 
-function toggleGuitarInput(event) {
-    var button = document.querySelector("#toggleGuitarIn");
-
-    if(!guitarPluggedIn) {
-        guitarInput.connect(amp.input);
-        button.innerHTML = "Guitar input: <span style='color:green;'>ACTIVATED</span>, click to toggle on/off!";
-        button.classList.remove("pulse");
-    } else {
-        guitarInput.disconnect();
-        button.innerHTML = "Guitar input: <span style='color:red;'>NOT ACTIVATED</span>, click to toggle on/off!";
-        button.classList.add("pulse");
-    }
-    guitarPluggedIn = !guitarPluggedIn;
-}
-
 // Visualizations
 var inputVisualization, outputVisualization;
 
 function initVisualizations() {
     inputVisualization = new Visualization();
-    inputVisualization.configure("inputSignalCanvas", analyzerAtInput);
+    inputVisualization.configure('inputSignalCanvas', analyzerAtInput);
 
     outputVisualization = new Visualization();
-    outputVisualization.configure("outputSignalCanvas", analyzerAtOutput);
+    outputVisualization.configure('outputSignalCanvas', analyzerAtOutput);
 
 
     // start updating the visualizations
@@ -123,7 +135,7 @@ function visualize() {
 }
 
 // effects
-//----------- EQUALIZER ----------- 
+// ----------- EQUALIZER -----------
 function Equalizer(ctx) {
     var filters = [];
 
@@ -132,20 +144,20 @@ function Equalizer(ctx) {
     [60, 170, 350, 1000, 3500, 10000].forEach(function (freq, i) {
         var eq = ctx.createBiquadFilter();
         eq.frequency.value = freq;
-        eq.type = "peaking";
+        eq.type = 'peaking';
         eq.gain.value = 0;
         filters.push(eq);
     });
 
     // Connect filters in serie
-    //sourceNode.connect(filters[0]);
+    // sourceNode.connect(filters[0]);
 
     for (var i = 0; i < filters.length - 1; i++) {
         filters[i].connect(filters[i + 1]);
     }
 
     // connect the last filter to the speakers
-    //filters[filters.length - 1].connect(ctx.destination);
+    // filters[filters.length - 1].connect(ctx.destination);
 
     function changeGain(sliderVal, nbFilter) {
         // sliderVal in [-30, +30]
@@ -153,16 +165,16 @@ function Equalizer(ctx) {
         filters[nbFilter].gain.value = value;
 
         // update output labels
-        //var output = document.querySelector("#gain" + nbFilter);
-        //output.value = value + " dB";
+        // var output = document.querySelector('#gain' + nbFilter);
+        // output.value = value + ' dB';
 
         // update sliders
-        //var numSlider = nbFilter + 1;
-        //var slider = document.querySelector("#EQ" + numSlider + "slider");
-        //slider.value = value;
+        // var numSlider = nbFilter + 1;
+        // var slider = document.querySelector('#EQ' + numSlider + 'slider');
+        // slider.value = value;
 
         // refresh amp slider state in the web component GUI
-        var sliderWC = document.querySelector("#slider" + (nbFilter+1));
+        var sliderWC = document.querySelector('#slider' + (nbFilter+1));
         // second parameter set to false will not fire an event
         sliderWC.setValue(parseFloat(sliderVal).toFixed(0), false);
     }
@@ -194,24 +206,24 @@ function Equalizer(ctx) {
 
 function Amp(context) {
     var presets = [];
-    var menuPresets = document.querySelector("#QFPresetMenu2");
-    var menuDisto1 = document.querySelector("#distorsionMenu1");
-    var menuDisto2 = document.querySelector("#distorsionMenu2");
+    var menuPresets = document.querySelector('#QFPresetMenu2');
+    var menuDisto1 = document.querySelector('#distorsionMenu1');
+    var menuDisto2 = document.querySelector('#distorsionMenu2');
     // for the waveshapers from the preamp
     var wsFactory = new WaveShapers();
     buildDistoMenu1();
     buildDistoMenu2();
 
-    var currentDistoName = "standard";
-    var currentK = 2; // we have separates ks, but also a "global" one that
+    var currentDistoName = 'standard';
+    var currentK = 2; // we have separates ks, but also a 'global' one that
                       // is the max of the two (the knob value)
     var currentWSCurve = wsFactory.distorsionCurves[currentDistoName](currentK);
     // for Wave Shaper Curves visualization
     var DRAWER_CANVAS_SIZE = 100;
-    var distoDrawer1 = new CurveDrawer("distoDrawerCanvas1");
-    var signalDrawer1 = new CurveDrawer("signalDrawerCanvas1");
-    var distoDrawer2 = new CurveDrawer("distoDrawerCanvas2");
-    var signalDrawer2 = new CurveDrawer("signalDrawerCanvas2");
+    var distoDrawer1 = new CurveDrawer('distoDrawerCanvas1');
+    var signalDrawer1 = new CurveDrawer('signalDrawerCanvas1');
+    var distoDrawer2 = new CurveDrawer('distoDrawerCanvas2');
+    var signalDrawer2 = new CurveDrawer('signalDrawerCanvas2');
 
     // ------------
     // PREAM STAGE
@@ -237,43 +249,11 @@ function Amp(context) {
     var od = [];
     var distoTypes = ['asymetric', 'standard'];
 
-    var gainsOds = [];
-
-    // Tonestack in serie, cf Lepou's mail  
-    /*
-    for (var i = 0; i < 4; i++) {
-        loCutFilters[i] = context.createBiquadFilter();
-        loCutFilters[i].type = "lowshelf";
-        loCutFilters[i].frequency.value = 720;
-        loCutFilters[i].gain.value = 3.3;
-
-        hiCutFilters[i] = context.createBiquadFilter();
-        hiCutFilters[i].type = "lowpass";
-        hiCutFilters[i].frequency.value = 12000;
-        hiCutFilters[i].Q.value = 0.7071;
-
-        highShelfBoosts[i] = context.createBiquadFilter();
-        highShelfBoosts[i].type = "highshelf";
-        highShelfBoosts[i].frequency.value = 12000; // Which values ?
-        highShelfBoosts[i].Q.value = 0.7071;        // Which values ?
-
-        od[i] = context.createWaveShaper();
-        od[i].curve = makeDistortionCurve(k[i]);
-        // Oversampling generates some (small) latency
-        //od[i].oversample = '4x';
-
-        // gains
-        gainsOds[i] = context.createGain();
-        gainsOds[i].gain.value = 1;
-    }
-
-    */
-   
     // JCM 800 preamp schematic...
     //
     // Low shelf cut -6db à 720Hz
     var lowShelf1 = context.createBiquadFilter();
-    lowShelf1.type = "lowshelf";
+    lowShelf1.type = 'lowshelf';
     lowShelf1.frequency.value = 720;
     lowShelf1.gain.value = -6;
 
@@ -281,7 +261,7 @@ function Amp(context) {
     // if vol = 50%, then filter at -6db, 320Hz
     // shoud go from -4db to -6db for +/- fatness
     var lowShelf2 = context.createBiquadFilter();
-    lowShelf2.type = "lowshelf";    
+    lowShelf2.type = 'lowshelf';
     lowShelf2.frequency.value = 320;
     lowShelf2.gain.value = -5;
 
@@ -289,22 +269,22 @@ function Amp(context) {
     var preampStage1Gain = context.createGain();
     preampStage1Gain.gain.value = 1.0;
 
-    // Distorsion 1, here we should use an asymetric function in order to 
+    // Distorsion 1, here we should use an asymetric function in order to
     // generate odd harmonics
     od[0] = context.createWaveShaper();
-    od[0].curve =  wsFactory.distorsionCurves[distoTypes[0]](0);
+    od[0].curve = wsFactory.distorsionCurves[distoTypes[0]](0);
     menuDisto1.value = distoTypes[0];
 
     // HighPass at 7-8 Hz, rectify the signal that got a DC value due
     // to the possible asymetric transfer function
     var highPass1 = context.createBiquadFilter();
-    highPass1.type = "highpass";    
+    highPass1.type = 'highpass';
     highPass1.frequency.value = 6;
     highPass1.Q.value = 0.7071;
 
     // lowshelf cut -6db 720Hz
     var lowShelf3 = context.createBiquadFilter();
-    lowShelf3.type = "lowshelf";    
+    lowShelf3.type = 'lowshelf';
     lowShelf3.frequency.value = 720;
     lowShelf3.gain.value = -6;
 
@@ -327,25 +307,24 @@ function Amp(context) {
     // ------------------------------
     // TONESTACK STAGES
     // ------------------------------
-    // Useless ?
-    var bassFilter = context.createBiquadFilter();
+    bassFilter = context.createBiquadFilter();
     bassFilter.frequency.value = 100;
-    bassFilter.type = "lowshelf";
+    bassFilter.type = 'lowshelf';
     bassFilter.Q.value = 0.7071; // To check with Lepou
 
-    var midFilter = context.createBiquadFilter();
+    midFilter = context.createBiquadFilter();
     midFilter.frequency.value = 1700;
-    midFilter.type = "peaking";
+    midFilter.type = 'peaking';
     midFilter.Q.value = 0.7071; // To check with Lepou
 
-    var trebleFilter = context.createBiquadFilter();
+    trebleFilter = context.createBiquadFilter();
     trebleFilter.frequency.value = 6500;
-    trebleFilter.type = "highshelf";
+    trebleFilter.type = 'highshelf';
     trebleFilter.Q.value = 0.7071; // To check with Lepou
 
-    var presenceFilter = context.createBiquadFilter();
+    presenceFilter = context.createBiquadFilter();
     presenceFilter.frequency.value = 3900;
-    presenceFilter.type = "peaking";
+    presenceFilter.type = 'peaking';
     presenceFilter.Q.value = 0.7071; // To check with Lepou
 
     // -----------------------------------
@@ -354,15 +333,14 @@ function Amp(context) {
     //  before EQ, filter highs and lows (Fred Miniere advise)
     var eqhicut = context.createBiquadFilter();
         eqhicut.frequency.value = 10000;
-        eqhicut.type = "peaking";
+        eqhicut.type = 'peaking';
         eqhicut.gain.value = -25;
 
     var eqlocut = context.createBiquadFilter();
         eqlocut.frequency.value = 60;
-        eqlocut.type = "peaking";
+        eqlocut.type = 'peaking';
         eqlocut.gain.value = -19;
 
-    
     var eq = new Equalizer(context);
     changeEQValues([0, 0, 0, 0, 0, 0]);
     var bypassEQg = context.createGain();
@@ -373,21 +351,9 @@ function Amp(context) {
     // Master volume
     var masterVolume = context.createGain();
     changeMasterVolume(2);
-/*
-    reverb = new Reverb(context, function () {
-        console.log("reverb created");
 
-        cabinetSim = new CabinetSimulator(context, function () {
-            console.log("cabinet sim created");
-
-            doAllConnections();
-
-        });
-    });
-*/
-
-    reverb = new Convolver(context, reverbImpulses, "reverbImpulses");
-    cabinetSim = new Convolver(context, cabinetImpulses, "cabinetImpulses");
+    reverb = new Convolver(context, Impulses.reverbImpulses, 'reverbImpulses');
+    cabinetSim = new Convolver(context, Impulses.cabinetImpulses, 'cabinetImpulses');
 
     doAllConnections();
 
@@ -405,19 +371,18 @@ function Amp(context) {
         initPresets();
 
         setDefaultPreset();
-        console.log("running");
+        console.log('running');
     }
-
 
     function buildGraph() {
         input.connect(inputGain);
         input.connect(byPass);
 
-        // boost is not activated, it's just a sort of disto at 
+        // boost is not activated, it's just a sort of disto at
         // the very beginning of the amp route
         inputGain.connect(boost.input);
         // JCM 800 like...
-        
+
         boost.output.connect(lowShelf1);
         lowShelf1.connect(lowShelf2);
         lowShelf2.connect(preampStage1Gain);
@@ -432,9 +397,8 @@ function Amp(context) {
 
         od[1].connect(outputGain);
 
-
         // tonestack
-        outputGain.connect(trebleFilter );
+        outputGain.connect(trebleFilter);
         trebleFilter.connect(bassFilter);
         bassFilter.connect(midFilter);
         midFilter.connect(presenceFilter);
@@ -442,7 +406,6 @@ function Amp(context) {
         // lo and hicuts
         presenceFilter.connect(eqlocut);
         eqlocut.connect(eqhicut);
-
 
         // post process
         eqhicut.connect(inputEQ);
@@ -452,22 +415,21 @@ function Amp(context) {
         bypassEQg.connect(masterVolume);
 
         // normal route
-        
         inputEQ.connect(eq.input);
         eq.output.connect(masterVolume);
         masterVolume.connect(reverb.input);
 
         reverb.output.connect(cabinetSim.input);
         cabinetSim.output.connect(output);
-        //eq.output.connect(output);
-        //reverb.output.connect(output);
+        // eq.output.connect(output);
+        // reverb.output.connect(output);
 
         // byPass route
         byPass.connect(output);
     }
 
-    function boostOnOff(cb) {  
-        // called when we click the switch on the GUI      
+    function boostOnOff(cb) {
+        // called when we click the switch on the GUI
         boost.toggle();
 
         adjustOutputGainIfBoostActivated();
@@ -475,40 +437,40 @@ function Amp(context) {
     }
 
     function changeBoost(state) {
-        console.log("changeBoost, boost before: " + boost.isActivated() + " output gain=" + output.gain.value );
+        console.log('changeBoost, boost before: ' + boost.isActivated() + ' output gain=' + output.gain.value );
 
         if(boost.isActivated() !== state) {
             // we need to adjust the output gain
-            console.log("changeBoost: we change boost state");
+            console.log('changeBoost: we change boost state');
             boost.onOff(state);
             adjustOutputGainIfBoostActivated();
             updateBoostLedButtonState(boost.isActivated());
         } else {
-            console.log("changeBoost: we do not change boost state");
+            console.log('changeBoost: we do not change boost state');
         }
 
-        console.log("changeBoost, boost after: " + boost.isActivated());
+        console.log('changeBoost, boost after: ' + boost.isActivated());
     }
 
     function adjustOutputGainIfBoostActivated() {
-        console.log("adjustOutputGainIfBoostActivated: output gain value before = " + output.gain.value)
+        console.log('adjustOutputGainIfBoostActivated: output gain value before = ' + output.gain.value)
 
         if(boost.isActivated()) {
             output.gain.value /= 2;
         } else {
             output.gain.value *= 2;
         }
-        console.log("adjustOutputGainIfBoostActivated: output gain value after = " + output.gain.value)
+        console.log('adjustOutputGainIfBoostActivated: output gain value after = ' + output.gain.value)
     }
 
     function updateBoostLedButtonState(activated) {
         // update buttons states
-        var boostSwitch = document.querySelector("#toggleBoost");
+        var boostSwitch = document.querySelector('#toggleBoost');
 
         if(boost.isActivated()) {
-            boostSwitch.setValue(1,false);
+            boostSwitch.setValue(1, false);
         } else {
-            boostSwitch.setValue(0,false);
+            boostSwitch.setValue(0, false);
         }
     }
 
@@ -519,7 +481,7 @@ function Amp(context) {
 
     function changeOutputGainValue(sliderVal) {
         output.gain.value = parseFloat(sliderVal)/10;
-        console.log("changeOutputGainValue value = " + output.gain.value);
+        console.log('changeOutputGainValue value = ' + output.gain.value);
     }
 
     // PREAMP
@@ -529,11 +491,11 @@ function Amp(context) {
         lowShelf1.frequency.value = value;
 
         // update output labels
-        var output = document.querySelector("#lowShelf1Freq");
-        output.value = parseFloat(sliderVal).toFixed(1) + " Hz";
+        var output = document.querySelector('#lowShelf1Freq');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' Hz';
 
         // refresh slider state
-        var slider = document.querySelector("#lowShelf1FreqSlider");
+        var slider = document.querySelector('#lowShelf1FreqSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -542,11 +504,11 @@ function Amp(context) {
         lowShelf1.gain.value = value;
 
         // update output labels
-        var output = document.querySelector("#lowShelf1Gain");
-        output.value = parseFloat(sliderVal).toFixed(1) + " dB";
+        var output = document.querySelector('#lowShelf1Gain');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' dB';
 
         // refresh slider state
-        var slider = document.querySelector("#lowShelf1GainSlider");
+        var slider = document.querySelector('#lowShelf1GainSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -555,11 +517,11 @@ function Amp(context) {
         lowShelf2.frequency.value = value;
 
         // update output labels
-        var output = document.querySelector("#lowShelf2Freq");
-        output.value = parseFloat(sliderVal).toFixed(1) + " Hz";
+        var output = document.querySelector('#lowShelf2Freq');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' Hz';
 
         // refresh slider state
-        var slider = document.querySelector("#lowShelf2FreqSlider");
+        var slider = document.querySelector('#lowShelf2FreqSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -567,13 +529,13 @@ function Amp(context) {
         var value = parseFloat(sliderVal);
         lowShelf2.gain.value = value;
 
-        console.log("lowshelf 2 gain = " + value);
+        console.log('lowshelf 2 gain = ' + value);
         // update output labels
-        var output = document.querySelector("#lowShelf2Gain");
-        output.value = parseFloat(sliderVal).toFixed(1) + " dB";
+        var output = document.querySelector('#lowShelf2Gain');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' dB';
 
         // refresh slider state
-        var slider = document.querySelector("#lowShelf2GainSlider");
+        var slider = document.querySelector('#lowShelf2GainSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -582,11 +544,11 @@ function Amp(context) {
         preampStage1Gain.gain.value = value;
 
         // update output labels
-        var output = document.querySelector("#preampStage1Gain");
+        var output = document.querySelector('#preampStage1Gain');
         output.value = parseFloat(sliderVal).toFixed(2);
 
         // refresh slider state
-        var slider = document.querySelector("#preampStage1GainSlider");
+        var slider = document.querySelector('#preampStage1GainSlider');
         slider.value = parseFloat(sliderVal).toFixed(2);
     }
 
@@ -595,11 +557,11 @@ function Amp(context) {
         highPass1.frequency.value = value;
 
         // update output labels
-        var output = document.querySelector("#highPass1Freq");
-        output.value = parseFloat(sliderVal).toFixed(1) + " Hz";
+        var output = document.querySelector('#highPass1Freq');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' Hz';
 
         // refresh slider state
-        var slider = document.querySelector("#highPass1FreqSlider");
+        var slider = document.querySelector('#highPass1FreqSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -608,11 +570,11 @@ function Amp(context) {
         highPass1.Q.value = value;
 
         // update output labels
-        var output = document.querySelector("#highPass1Q");
+        var output = document.querySelector('#highPass1Q');
         output.value = parseFloat(sliderVal).toFixed(4);
 
         // refresh slider state
-        var slider = document.querySelector("#highPass1QSlider");
+        var slider = document.querySelector('#highPass1QSlider');
         slider.value = parseFloat(sliderVal).toFixed(4);
     }
 
@@ -621,11 +583,11 @@ function Amp(context) {
         lowShelf3.frequency.value = value;
 
         // update output labels
-        var output = document.querySelector("#lowShelf3Freq");
-        output.value = parseFloat(sliderVal).toFixed(1) + " Hz";
+        var output = document.querySelector('#lowShelf3Freq');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' Hz';
 
         // refresh slider state
-        var slider = document.querySelector("#lowShelf3FreqSlider");
+        var slider = document.querySelector('#lowShelf3FreqSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -634,11 +596,11 @@ function Amp(context) {
         lowShelf3.gain.value = value;
 
         // update output labels
-        var output = document.querySelector("#lowShelf3Gain");
-        output.value = parseFloat(sliderVal).toFixed(1) + " dB";
+        var output = document.querySelector('#lowShelf3Gain');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' dB';
 
         // refresh slider state
-        var slider = document.querySelector("#lowShelf3GainSlider");
+        var slider = document.querySelector('#lowShelf3GainSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -647,11 +609,11 @@ function Amp(context) {
         preampStage2Gain.gain.value = value;
 
         // update output labels
-        var output = document.querySelector("#preampStage2Gain");
+        var output = document.querySelector('#preampStage2Gain');
         output.value = parseFloat(sliderVal).toFixed(2);
 
         // refresh slider state
-        var slider = document.querySelector("#preampStage2GainSlider");
+        var slider = document.querySelector('#preampStage2GainSlider');
         slider.value = parseFloat(sliderVal).toFixed(2);
     }
 
@@ -663,11 +625,11 @@ function Amp(context) {
             hiCutFilters[i].frequency.value = value;
     }
         // update output labels
-        var output = document.querySelector("#hiCutFreq");
-        output.value = parseFloat(sliderVal).toFixed(1) + " Hz";
+        var output = document.querySelector('#hiCutFreq');
+        output.value = parseFloat(sliderVal).toFixed(1) + ' Hz';
 
         // refresh slider state
-        var slider = document.querySelector("#hiCutFreqSlider");
+        var slider = document.querySelector('#hiCutFreqSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
     }
 
@@ -675,19 +637,19 @@ function Amp(context) {
         // sliderVal is in [0, 10]
         var value = parseFloat(sliderVal);
         bassFilter.gain.value = (value-10) * 7;
-        console.log("bass gain set to " + bassFilter.gain.value);
+        console.log('bass gain set to ' + bassFilter.gain.value);
 
         // update output labels
-        //var output = document.querySelector("#bassFreq");
+        //var output = document.querySelector('#bassFreq');
         //output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#bassFreqSlider");
+        //var slider = document.querySelector('#bassFreqSlider');
         //slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
         //sliderVal = value / 7 + 10;
-        var knob = document.querySelector("#Knob4");
+        var knob = document.querySelector('#Knob4');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -697,16 +659,16 @@ function Amp(context) {
         midFilter.gain.value = (value-5) * 4;
 
         // update output labels
-        //var output = document.querySelector("#midFreq");
+        //var output = document.querySelector('#midFreq');
         //output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#midFreqSlider");
+        //var slider = document.querySelector('#midFreqSlider');
         //slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
          //sliderVal = value /4 + 5;
-        var knob = document.querySelector("#Knob5");
+        var knob = document.querySelector('#Knob5');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -716,16 +678,16 @@ function Amp(context) {
         trebleFilter.gain.value = (value-10) * 10;
 
         // update output labels
-        //var output = document.querySelector("#trebleFreq");
+        //var output = document.querySelector('#trebleFreq');
         //output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#trebleFreqSlider");
+        //var slider = document.querySelector('#trebleFreqSlider');
         //slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
         //sliderVal = value /10 + 10;
-        var knob = document.querySelector("#Knob6");
+        var knob = document.querySelector('#Knob6');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -733,52 +695,52 @@ function Amp(context) {
         // sliderVal is in [0, 10]
         var value = parseFloat(sliderVal);
         presenceFilter.gain.value = (value-5) * 2;
-        //console.log("set presence freq to " + presenceFilter.frequency.value)
+        //console.log('set presence freq to ' + presenceFilter.frequency.value)
 
         // update output labels
-        //var output = document.querySelector("#presenceFreq");
+        //var output = document.querySelector('#presenceFreq');
         //output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#presenceFreqSlider");
+        //var slider = document.querySelector('#presenceFreqSlider');
         //slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
-        var knob = document.querySelector("#Knob8");
+        var knob = document.querySelector('#Knob8');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
     // Build a drop down menu with all distorsion names
     function buildDistoMenu1() {
         for(var p in wsFactory.distorsionCurves) {
-            var option = document.createElement("option");
+            var option = document.createElement('option');
             option.value = p;
             option.text = p;
-            menuDisto1.appendChild(option);    
+            menuDisto1.appendChild(option);
         }
         menuDisto1.onchange = changeDistoType1;
     }
     // Build a drop down menu with all distorsion names
     function buildDistoMenu2() {
         for(var p in wsFactory.distorsionCurves) {
-            var option = document.createElement("option");
+            var option = document.createElement('option');
             option.value = p;
             option.text = p;
-            menuDisto2.appendChild(option);    
+            menuDisto2.appendChild(option);
         }
         menuDisto2.onchange = changeDistoType2;
     }
 
     function changeDistoType1() {
-        console.log("Changing disto1 to : " + menuDisto1.value);
-        currentDistoName = menuDisto1.value;   
+        console.log('Changing disto1 to : ' + menuDisto1.value);
+        currentDistoName = menuDisto1.value;
         distoTypes[0] = currentDistoName;
         changeDrive(currentK);
     }
 
     function changeDistoType2() {
-        console.log("Changing disto2 to : " + menuDisto2.value);
-        currentDistoName = menuDisto2.value;   
+        console.log('Changing disto2 to : ' + menuDisto2.value);
+        currentDistoName = menuDisto2.value;
         distoTypes[1] = currentDistoName;
         changeDrive(currentK);
     }
@@ -799,19 +761,19 @@ function Amp(context) {
 
     function changeDrive(sliderValue) {
       // sliderValue in [0,10]
-      // We can imagine having some "profiles here" -> generate
+      // We can imagine having some 'profiles here' -> generate
       // different K values from one single sliderValue for the
       // drive.
       // other values i.e [0.5, 0.5, 0.8, 1] -> less distorsion
       // on bass frequencies and top high frequency
-      
+
       for(var i = 0; i < 2; i++) {
             changeDistorsionValues(sliderValue, i);
       }
     }
 
     function changeDistorsionValues(sliderValue, numDisto) {
-        // sliderValue is in [0, 10] range, adjust to [0, 1500] range  
+        // sliderValue is in [0, 10] range, adjust to [0, 1500] range
         var value = 150 * parseFloat(sliderValue);
         var minp = 0;
         var maxp = 1500;
@@ -827,24 +789,24 @@ function Amp(context) {
         // end of logarithmic adjustment
 
         k[numDisto] = value;
-        //console.log("k = " + value + " pos = " + logToPos(value));
-        //console.log("distoTypes = " + distoTypes[numDisto]);
+        //console.log('k = ' + value + ' pos = ' + logToPos(value));
+        //console.log('distoTypes = ' + distoTypes[numDisto]);
         od[numDisto].curve = wsFactory.distorsionCurves[distoTypes[numDisto]](k[numDisto]);//makeDistortionCurve(k[numDisto]);
         currentWSCurve = od[numDisto].curve;
         //od[numDisto].curve = makeDistortionCurve(sliderValue);
         //makeDistortionCurve(k[numDisto]);
         //od[numDisto].curve = makeDistortionCurve(sliderValue);
         // update output labels
-        var output = document.querySelector("#k" + numDisto);
+        var output = document.querySelector('#k' + numDisto);
         output.value = parseFloat(sliderValue).toFixed(1);
 
         // update sliders
         var numSlider = numDisto + 1;
-        var slider = document.querySelector("#K" + numSlider + "slider");
+        var slider = document.querySelector('#K' + numSlider + 'slider');
         slider.value = parseFloat(sliderValue).toFixed(1);
 
         // refresh knob state
-        var knob = document.querySelector("#Knob3");
+        var knob = document.querySelector('#Knob3');
         var maxPosVal1 = Math.max(logToPos(k[2]), logToPos(k[3]));
         var maxPosVal2 = Math.max(logToPos(k[0]), logToPos(k[1]));
         var maxPosVal = Math.max(maxPosVal1, maxPosVal2);
@@ -873,22 +835,22 @@ function Amp(context) {
 
     function changeOversampling(cb) {
         for (var i = 0; i < 2; i++) {
-            if(cb.checked) {
+            if (cb.checked) {
                 // Oversampling generates some (small) latency
                 od[i].oversample = '4x';
                 boost.setOversampling('4x');
-                console.log("set oversampling to 4x");
+                console.log('set oversampling to 4x');
             } else {
                 od[i].oversample = 'none';
                  boost.setOversampling('none');
-                console.log("set oversampling to none");
+                console.log('set oversampling to none');
             }
          }
          // Not sure if this is necessary... My ears can't hear the difference
          // between oversampling=node and 4x ? Maybe we should re-init the
          // waveshaper curves ?
-         //changeDistoType1();
-         //changeDistoType2();
+         // changeDistoType1();
+         // changeDistoType2();
     }
 
     // Returns an array of distorsions values in [0, 10] range
@@ -951,12 +913,12 @@ function Amp(context) {
         filters[numQ].Q.value = value;
 
         // update output labels
-        var output = document.querySelector("#q" + numQ);
+        var output = document.querySelector('#q' + numQ);
         output.value = value.toFixed(1);
 
         // update sliders
         var numSlider = numQ + 1;
-        var slider = document.querySelector("#Q" + numSlider + "slider");
+        var slider = document.querySelector('#Q' + numSlider + 'slider');
         slider.value = value;
 
     }
@@ -966,11 +928,11 @@ function Amp(context) {
         filters[numF].frequency.value = value;
 
         // update output labels
-        var output = document.querySelector("#freq" + numF);
-        output.value = value + " Hz";
+        var output = document.querySelector('#freq' + numF);
+        output.value = value + ' Hz';
         // refresh slider state
         var numSlider = numF + 1;
-        var slider = document.querySelector("#F" + numSlider + "slider");
+        var slider = document.querySelector('#F' + numSlider + 'slider');
         slider.value = value;
     }
 
@@ -982,15 +944,15 @@ function Amp(context) {
         outputGain.gain.value = value;
 
         // update output labels
-        //var output = document.querySelector("#outputGain");
-        //output.value = parseFloat(sliderVal).toFixed(1);
+        // var output = document.querySelector('#outputGain');
+        // output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#OGslider");
-        //slider.value = parseFloat(sliderVal).toFixed(1);
+        // var slider = document.querySelector('#OGslider');
+        // slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
-        var knob = document.querySelector("#Knob1");
+        var knob = document.querySelector('#Knob1');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -1002,15 +964,15 @@ function Amp(context) {
         inputGain.gain.value = value;
 
         // update output labels
-        //var output = document.querySelector("#outputGain");
-        //output.value = parseFloat(sliderVal).toFixed(1);
+        // var output = document.querySelector('#outputGain');
+        // output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#OGslider");
-        //slider.value = parseFloat(sliderVal).toFixed(1);
+        // var slider = document.querySelector('#OGslider');
+        // slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
-        var knob = document.querySelector("#Knob1");
+        var knob = document.querySelector('#Knob1');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -1020,15 +982,15 @@ function Amp(context) {
         masterVolume.gain.value = value;
 
         // update output labels
-        //var output = document.querySelector("#MVOutputGain");
+        //var output = document.querySelector('#MVOutputGain');
         //output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#MVslider");
+        //var slider = document.querySelector('#MVslider');
         //slider.value = parseFloat(sliderVal).toFixed(1);
-        
+
         // refresh knob state
-        var knob = document.querySelector("#Knob2");
+        var knob = document.querySelector('#Knob2');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -1039,15 +1001,15 @@ function Amp(context) {
         reverb.setGain(value);
 
         // update output labels
-        //var output = document.querySelector("#reverbGainOutput");
+        //var output = document.querySelector('#reverbGainOutput');
         //output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        //var slider = document.querySelector("#convolverSlider");
+        //var slider = document.querySelector('#convolverSlider');
         //slider.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh knob state
-        var knob = document.querySelector("#Knob7");
+        var knob = document.querySelector('#Knob7');
         knob.setValue(parseFloat(sliderVal).toFixed(1), false);
     }
 
@@ -1063,11 +1025,11 @@ function Amp(context) {
         cabinetSim.setGain(value);
 
         // update output labels
-        var output = document.querySelector("#cabinetGainOutput");
+        var output = document.querySelector('#cabinetGainOutput');
         output.value = parseFloat(sliderVal).toFixed(1);
 
         // refresh slider state
-        var slider = document.querySelector("#convolverCabinetSlider");
+        var slider = document.querySelector('#convolverCabinetSlider');
         slider.value = parseFloat(sliderVal).toFixed(1);
 
     }
@@ -1091,34 +1053,34 @@ function Amp(context) {
     // --------
     function initPresets() {
         // updated 10/4/2016
-        var preset1 = {"name":"Hard Rock classic 1","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-5,"gain1":1,"distoName1":"asymetric","K1":"7.8","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"notSoDistorded","K2":"7.8","OG":"7.0","BF":"8.2","MF":"8.2","TF":"3.8","PF":"6.9","EQ":[5,11,-6,-10,7,2],"MV":"7.2","RN":"Fender Hot Rod","RG":"2.0","CN":"Marshall 1960, axis","CG":"9.4"};
+        var preset1 = {'name':'Hard Rock classic 1','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-5,'gain1':1,'distoName1':'asymetric','K1':'7.8','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'notSoDistorded','K2':'7.8','OG':'7.0','BF':'8.2','MF':'8.2','TF':'3.8','PF':'6.9','EQ':[5,11,-6,-10,7,2],'MV':'7.2','RN':'Fender Hot Rod','RG':'2.0','CN':'Marshall 1960, axis','CG':'9.4'};
         presets.push(preset1);
 
-        var preset2 = {"name":"Clean and Warm","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":1.600000023841858,"gain1":1,"distoName1":"asymetric","K1":"7.8","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"standard","K2":"0.9","OG":"7.0","BF":"6.7","MF":"7.1","TF":"3.2","PF":"6.9","EQ":[10,5,-7,-7,16,0],"MV":"7.2","RN":"Fender Hot Rod","RG":"1.4","CN":"Marshall 1960, axis","CG":"8.8"};
+        var preset2 = {'name':'Clean and Warm','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':1.600000023841858,'gain1':1,'distoName1':'asymetric','K1':'7.8','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'standard','K2':'0.9','OG':'7.0','BF':'6.7','MF':'7.1','TF':'3.2','PF':'6.9','EQ':[10,5,-7,-7,16,0],'MV':'7.2','RN':'Fender Hot Rod','RG':'1.4','CN':'Marshall 1960, axis','CG':'8.8'};
         presets.push(preset2);
 
-        var preset3 = {"name":"Strong and Warm","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-1,"gain1":1.0299999713897705,"distoName1":"asymetric","K1":"7.8","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"superClean","K2":"7.8","OG":"7.0","BF":"8.2","MF":"6.7","TF":"5.0","PF":"6.9","EQ":[0,0,0,-1,0,1],"MV":"5.9","RN":"Fender Hot Rod","RG":"1.1","CN":"Vox Custom Bright 4x12 M930 Axis 1","CG":"8.0"};
+        var preset3 = {'name':'Strong and Warm','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-1,'gain1':1.0299999713897705,'distoName1':'asymetric','K1':'7.8','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'superClean','K2':'7.8','OG':'7.0','BF':'8.2','MF':'6.7','TF':'5.0','PF':'6.9','EQ':[0,0,0,-1,0,1],'MV':'5.9','RN':'Fender Hot Rod','RG':'1.1','CN':'Vox Custom Bright 4x12 M930 Axis 1','CG':'8.0'};
         presets.push(preset3);
 
-        //preset4 = {"name":"Fat sound","boost":true,"LS1Freq":720,"LS1Gain":-5.800000190734863,"LS2Freq":320,"LS2Gain":6.599999904632568,"gain1":0.11999999731779099,"distoName1":"asymetric","K1":"5.4","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-5.199999809265137,"gain2":1,"distoName2":"standard","K2":"5.4","OG":"3.5","BF":"3.2","MF":"5.0","TF":"5.0","PF":"9.7","EQ":[1,0,-6,-8,-6,-30],"MV":"3.1","RN":"Fender Hot Rod","RG":"0.0","CN":"Marshall 1960, axis","CG":"3.4"};
+        //preset4 = {'name':'Fat sound','boost':true,'LS1Freq':720,'LS1Gain':-5.800000190734863,'LS2Freq':320,'LS2Gain':6.599999904632568,'gain1':0.11999999731779099,'distoName1':'asymetric','K1':'5.4','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-5.199999809265137,'gain2':1,'distoName2':'standard','K2':'5.4','OG':'3.5','BF':'3.2','MF':'5.0','TF':'5.0','PF':'9.7','EQ':[1,0,-6,-8,-6,-30],'MV':'3.1','RN':'Fender Hot Rod','RG':'0.0','CN':'Marshall 1960, axis','CG':'3.4'};
         //presets.push(preset4);
-        var preset4 = {"name":"Clean no reverb","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-6.300000190734863,"gain1":1,"distoName1":"asymetric","K1":"2.1","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"crunch","K2":"2.1","OG":"7.0","BF":"6.7","MF":"5.0","TF":"5.0","PF":"8.9","EQ":[4,13,-8,-8,15,12],"MV":"3.7","RN":"Fender Hot Rod","RG":"0.0","CN":"Marshall 1960, axis","CG":"4.5"};
+        var preset4 = {'name':'Clean no reverb','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-6.300000190734863,'gain1':1,'distoName1':'asymetric','K1':'2.1','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'crunch','K2':'2.1','OG':'7.0','BF':'6.7','MF':'5.0','TF':'5.0','PF':'8.9','EQ':[4,13,-8,-8,15,12],'MV':'3.7','RN':'Fender Hot Rod','RG':'0.0','CN':'Marshall 1960, axis','CG':'4.5'};
         presets.push(preset4);
 
-        var preset5 = {"name":"Another Clean Sound","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-6.300000190734863,"gain1":1,"distoName1":"asymetric","K1":"6.4","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"crunch","K2":"6.4","OG":"7.0","BF":"6.7","MF":"5.0","TF":"5.0","PF":"8.9","EQ":[4,13,-8,-8,15,12],"MV":"3.7","RN":"Fender Hot Rod","RG":"2","CN":"Marshall 1960, axis","CG":"4.5"};
+        var preset5 = {'name':'Another Clean Sound','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-6.300000190734863,'gain1':1,'distoName1':'asymetric','K1':'6.4','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'crunch','K2':'6.4','OG':'7.0','BF':'6.7','MF':'5.0','TF':'5.0','PF':'8.9','EQ':[4,13,-8,-8,15,12],'MV':'3.7','RN':'Fender Hot Rod','RG':'2','CN':'Marshall 1960, axis','CG':'4.5'};
         presets.push(preset5);
 
-        var preset6 = {"name":"Mostly even harmonics","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-7.5,"gain1":1,"distoName1":"standard","K1":"6.7","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"standard","K2":"6.7","OG":"7.0","BF":"4.3","MF":"2.6","TF":"6.1","PF":"4.2","EQ":[5,12,-5,-10,2,10],"MV":"1.7","RN":"Fender Hot Rod","RG":"0.0","CN":"Vintage Marshall 1","CG":"8.4"};
+        var preset6 = {'name':'Mostly even harmonics','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-7.5,'gain1':1,'distoName1':'standard','K1':'6.7','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'standard','K2':'6.7','OG':'7.0','BF':'4.3','MF':'2.6','TF':'6.1','PF':'4.2','EQ':[5,12,-5,-10,2,10],'MV':'1.7','RN':'Fender Hot Rod','RG':'0.0','CN':'Vintage Marshall 1','CG':'8.4'};
         presets.push(preset6);
 
-        var preset7 = {"name":"Hard Rock classic 2","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-10.199999809265137,"gain1":1,"distoName1":"standard","K1":"5.2","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"notSoDistorded","K2":"5.1","OG":"7.0","BF":"8.7","MF":"8.0","TF":"3.8","PF":"9.4","EQ":[19,8,-6,-10,7,2],"MV":"5.5","RN":"Fender Hot Rod","RG":"0.7","CN":"Marshall 1960, axis","CG":"9.2"};
+        var preset7 = {'name':'Hard Rock classic 2','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-10.199999809265137,'gain1':1,'distoName1':'standard','K1':'5.2','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'notSoDistorded','K2':'5.1','OG':'7.0','BF':'8.7','MF':'8.0','TF':'3.8','PF':'9.4','EQ':[19,8,-6,-10,7,2],'MV':'5.5','RN':'Fender Hot Rod','RG':'0.7','CN':'Marshall 1960, axis','CG':'9.2'};
         presets.push(preset7);
-        
-        var preset8 = {"name":"SuperClean/Jazz","boost":false,"LS1Freq":720,"LS1Gain":-6,"LS2Freq":320,"LS2Gain":-6.300000190734863,"gain1":1,"distoName1":"crunch","K1":"5.4","HP1Freq":6,"HP1Q":0.707099974155426,"LS3Freq":720,"LS3Gain":-6,"gain2":1,"distoName2":"crunch","K2":"5.4","OG":"7.0","BF":"7.0","MF":"5.1","TF":"5.2","PF":"3.1","EQ":[10,7,0,-10,5,12],"MV":"3.8","RN":"Fender Hot Rod","RG":"1.5","CN":"Marshall 1960, axis","CG":"4.5"};
+
+        var preset8 = {'name':'SuperClean/Jazz','boost':false,'LS1Freq':720,'LS1Gain':-6,'LS2Freq':320,'LS2Gain':-6.300000190734863,'gain1':1,'distoName1':'crunch','K1':'5.4','HP1Freq':6,'HP1Q':0.707099974155426,'LS3Freq':720,'LS3Gain':-6,'gain2':1,'distoName2':'crunch','K2':'5.4','OG':'7.0','BF':'7.0','MF':'5.1','TF':'5.2','PF':'3.1','EQ':[10,7,0,-10,5,12],'MV':'3.8','RN':'Fender Hot Rod','RG':'1.5','CN':'Marshall 1960, axis','CG':'4.5'};
         presets.push(preset8);
 
         presets.forEach(function (p, index) {
-            var option = document.createElement("option");
+            var option = document.createElement('option');
             option.value = index;
             option.text = p.name;
             menuPresets.appendChild(option);
@@ -1132,10 +1094,10 @@ function Amp(context) {
 
     function setPreset(p) {
         if(p.distoName1 === undefined) {
-            p.distoName1 = "standard";
+            p.distoName1 = 'standard';
         }
          if(p.distoName2 === undefined) {
-            p.distoName2 = "standard";
+            p.distoName2 = 'standard';
         }
 
         if(p.boost === undefined) p.boost = false;
@@ -1186,7 +1148,7 @@ function Amp(context) {
     function printCurrentAmpValues() {
         var currentPresetValue = {
             name: 'current',
-            
+
             boost: boost.isActivated(),
 
             LS1Freq: lowShelf1.frequency.value,
@@ -1224,7 +1186,7 @@ function Amp(context) {
     // END PRESETS
 
     function bypass(cb) {
-        console.log("byPass : " + cb.checked);
+        console.log('byPass : ' + cb.checked);
 
         if (cb.checked) {
             // byPass mode
@@ -1237,11 +1199,11 @@ function Amp(context) {
         }
 
         // update buttons states
-        //var onOffButton = document.querySelector("#myonoffswitch");
-        var led = document.querySelector("#led");
+        //var onOffButton = document.querySelector('#myonoffswitch');
+        var led = document.querySelector('#led');
 
         //onOffButton.checked = cb.checked;
-        var onOffSwitch = document.querySelector("#switch1");
+        var onOffSwitch = document.querySelector('#switch1');
         if(cb.checked) {
             onOffSwitch.setValue(0,false);
             led.setValue(1, false);
@@ -1252,7 +1214,7 @@ function Amp(context) {
     }
 
     function bypassEQ(cb) {
-        console.log("EQ byPass : " + cb.checked);
+        console.log('EQ byPass : ' + cb.checked);
 
         if (cb.checked) {
             // byPass mode
@@ -1265,11 +1227,11 @@ function Amp(context) {
         }
 
         // update buttons states
-        //var onOffButton = document.querySelector("#myonoffswitch");
-        var led = document.querySelector("#led");
+        //var onOffButton = document.querySelector('#myonoffswitch');
+        var led = document.querySelector('#led');
 
         //onOffButton.checked = cb.checked;
-        var eqOnOffSwitch = document.querySelector("#switch2");
+        var eqOnOffSwitch = document.querySelector('#switch2');
         if(cb.checked) {
             eqOnOffSwitch.setValue(0,false);
         } else {
@@ -1278,7 +1240,7 @@ function Amp(context) {
     }
 
     // API: methods exposed
-    return { 
+    return {
         input: input,
         output: output,
         boostOnOff:boostOnOff,
@@ -1318,169 +1280,13 @@ function Amp(context) {
         setPreset: setPreset,
         printCurrentAmpValues : printCurrentAmpValues,
         bypass: bypass,
-        bypassEQ: bypassEQ
+        bypassEQ: bypassEQ,
+
+        toggleGuitarInput: toggleGuitarInput
     };
 }
 
-var reverbImpulses = [
-        {
-            name: "Fender Hot Rod",
-            url: "assets/impulses/reverb/cardiod-rear-levelled.wav"
-        },
-        {
-            name: "PCM 90 clean plate",
-            url: "assets/impulses/reverb/pcm90cleanplate.wav"
-        },
-        {
-            name: "Scala de Milan",
-            url: "assets/impulses/reverb/ScalaMilanOperaHall.wav"
-        }
-    ];
-var cabinetImpulses = [
-{
-            name: "Marshall 1960, axis",
-            url: "assets/impulses/cabinet/Marshall1960.wav"
-        },    
-        {
-            name: "Vintage Marshall 1",
-            url: "assets/impulses/cabinet/Block%20Inside.wav"
-        },
-        {
-            name: "Vox Custom Bright 4x12 M930 Axis 1",
-            url: "assets/impulses/cabinet/voxCustomBrightM930OnAxis1.wav"
-        },
-        {
-            name: "Fender Champ, axis",
-            url: "assets/impulses/cabinet/FenderChampAxisStereo.wav"
-        }
-    ];
-// ------- CONVOLVER, used for both reverb and cabinet simulation -------------------
-function Convolver(context, impulses, menuId) {
-    var convolverNode, convolverGain, directGain;
-    // create source and gain node
-    var inputGain = context.createGain();
-    var outputGain = context.createGain();
-    var decodedImpulse;
-
-    var menuIRs;
-    var IRs = impulses;
-
-    var currentImpulse = IRs[0];
-    var defaultImpulseURL = IRs[0].url;
-
-    convolverNode = context.createConvolver();
-    convolverNode.buffer = decodedImpulse;
-
-    convolverGain = context.createGain();
-    convolverGain.gain.value = 0;
-
-    directGain = context.createGain();
-    directGain.gain.value = 1;
-
-    buildIRsMenu(menuId);
-    buildAudioGraphConvolver();
-    setGain(0.2);
-    loadImpulseByUrl(defaultImpulseURL);
-    
-
-    function loadImpulseByUrl(url) {
-        // Load default impulse
-        const samples = Promise.all([loadSample(context,url)]).then(setImpulse);
-    }
-
-    function loadImpulseByName(name) {
-        if(name === undefined) {
-            name = IRs[0].name;
-            console.log("loadImpulseByName: name undefined, loading default impulse " + name);
-        }
-
-        var url="none";
-        // get url corresponding to name
-        for(var i=0; i < IRs.length; i++) {
-            if(IRs[i].name === name) {
-                url = IRs[i].url;
-                currentImpulse = IRs[i];
-                menuIRs.value = i;
-                break;
-            }
-        }
-        if(url === "none") {
-            console.log("ERROR loading reverb impulse name = " + name);
-        } else {
-            console.log("loadImpulseByName loading " + currentImpulse.name);
-            loadImpulseByUrl(url);
-        }
-    }
-
-    function loadImpulseFromMenu() {
-        var url = IRs[menuIRs.value].url;
-        currentImpulse = IRs[menuIRs.value];
-        console.log("loadImpulseFromMenu loading " + currentImpulse.name);
-        loadImpulseByUrl(url);
-    }
-
-    function setImpulse(param) {
-     // we get here only when the impulse is loaded and decoded
-        console.log("impulse loaded and decoded");
-        convolverNode.buffer = param[0];
-        console.log("convolverNode.buffer set with the new impulse (loaded and decoded");
-    }
-
-    function buildAudioGraphConvolver() {
-        // direct/dry route source -> directGain -> destination
-        inputGain.connect(directGain);
-        directGain.connect(outputGain);
-
-        // wet route with convolver: source -> convolver 
-        // -> convolverGain -> destination
-        inputGain.connect(convolverNode);
-        convolverNode.connect(convolverGain);
-        convolverGain.connect(outputGain);
-    }
-
-    function setGain(value) {
-        var v1 = Math.cos(value * Math.PI / 2);
-        var v2 = Math.cos((1 - value) * Math.PI / 2);
-
-        directGain.gain.value = v1;
-        convolverGain.gain.value = v2;
-    }
-
-    function getGain() {
-        return 2 * Math.acos(directGain.gain.value) / Math.PI;
-    }
-
-    function getName() {
-        return currentImpulse.name;
-    }
-
-
-    function buildIRsMenu(menuId) {
-        menuIRs = document.querySelector("#" + menuId);
-
-        IRs.forEach(function (impulse, index) {
-            var option = document.createElement("option");
-            option.value = index;
-            option.text = impulse.name;
-            menuIRs.appendChild(option);
-        });
-
-        menuIRs.oninput = loadImpulseFromMenu;
-    }
-    //--------------------------------------
-    // API : exposed methods and properties
-    // -------------------------------------
-    return {
-        input: inputGain,
-        output: outputGain,
-        setGain: setGain,
-        getGain: getGain,
-        getName:getName,
-        loadImpulseByName: loadImpulseByName
-    };
-}
-
-// Booster, useful to add a "Boost channel"
+// Booster, useful to add a 'Boost channel'
 var Boost = function(context) {
     // Booster not activated by default
     var activated = false;
@@ -1539,7 +1345,7 @@ var Boost = function(context) {
 
     function setOversampling(value) {
         shaper.oversample = value;
-        console.log("boost set oversampling to " + value);
+        console.log('boost set oversampling to ' + value);
     }
 
     function makeDistortionCurve(k) {
@@ -1562,4 +1368,3 @@ var Boost = function(context) {
         setOversampling: setOversampling
     };
 };
-
