+++ js/amp/impulses.js
@@ -1,0 +1,36 @@
+
+var Impulses = {};
+Impulses.reverbImpulses = [
+  {
+      name: 'Fender Hot Rod',
+      url: 'assets/impulses/reverb/cardiod-rear-levelled.wav'
+  },
+  {
+      name: 'PCM 90 clean plate',
+      url: 'assets/impulses/reverb/pcm90cleanplate.wav'
+  },
+  {
+      name: 'Scala de Milan',
+      url: 'assets/impulses/reverb/ScalaMilanOperaHall.wav'
+  }
+];
+Impulses.cabinetImpulses = [
+  {
+      name: 'Marshall 1960, axis',
+      url: 'assets/impulses/cabinet/Marshall1960.wav'
+  },
+  {
+      name: 'Vintage Marshall 1',
+      url: 'assets/impulses/cabinet/Block%20Inside.wav'
+  },
+  {
+      name: 'Vox Custom Bright 4x12 M930 Axis 1',
+      url: 'assets/impulses/cabinet/voxCustomBrightM930OnAxis1.wav'
+  },
+  {
+      name: 'Fender Champ, axis',
+      url: 'assets/impulses/cabinet/FenderChampAxisStereo.wav'
+  }
+];
+
+export default Impulses;
